+++ src/instrumentation/metamodel/llamaindex/entities/retrieval.js
@@ -1,46 +1,42 @@
-const { extractAssistantMessage } = require("../utils.js")   
+const { extractAssistantMessage } = require("../../utils")   
 
 exports.config = {
-    "type": "inference",
+    "type": "retrieval",
     "attributes": [
         [
             {
-                "_comment": "provider type ,name , deployment , inference_endpoint",
-                "attribute": "type",
+                "_comment": "vector store name",
+                "attribute": "name",
+                "accessor": function ({ instance, args }) {
+                    return instance?.index?.vectorStores?.TEXT?.constructor.name
+                }
+            },
+            {
+                "attribute": "vector store type",
                 "accessor": function ({ instance }) {
-                    if (instance?.constructor?.name?.toLowerCase().includes("azurechatopenai")) {
-                        return "inference.azure_openai"
-                    }
-                    if (instance?.constructor?.name?.toLowerCase().includes("chatopenai")) {
-                        return "inference.openai"
-                    }
+                    return "vectorstore." + instance?.index?.vectorStores?.TEXT?.constructor.name
                 }
             },
             {
                 "attribute": "deployment",
-                "accessor": function ({ instance }) {
-                    return instance.engine || instance.deployment || instance.deployment_name || instance.deployment_id || instance.azure_deployment
-                }
-            },
-            {
-                "attribute": "inference_endpoint",
-                "accessor": function ({ instance }) {
-                    return instance.azure_endpoint || instance.api_base || instance?.client?.baseURL
+                "accessor": function ({ instance, args }) {
+                    return ""
                 }
             }
         ],
         [
             {
-                "_comment": "LLM Model",
+                "_comment": "Embedding model name",
                 "attribute": "name",
                 "accessor": function ({ instance }) {
-                    return instance.model_name || instance.model
+                    return instance.index.vectorStores.TEXT.embedModel.model
                 }
             },
             {
+                _comment: "Embedding model type",
                 "attribute": "type",
                 "accessor": function ({ instance }) {
-                    return "model.llm" + (instance.model_name || instance.model)
+                    return "model.embedding." + instance.index.vectorStores.TEXT.embedModel.model
                 }
             }
         ]
