+++ src/instrumentation/metamodel/openai/entities/inference.js
